Moved 'helper' files back into a helpers directory, but inside the libs.
