+++ test/routes/shelves.test.js
@@ -24,7 +24,7 @@ const {
     recognize200,
     recognize400,
     recognize404,
-} = require('../../helpers/mocha/assert'); // Helper Mocha Assert Tests
+} = require('../../libs/helpers/mocha/assert'); // Helper Mocha Assert Tests
 
 // Global Variables
 let mongoServer;
